Cleaned up how vulkan is handled. A lot!
