--- VulkanCore/src/VulkanCore/Core/API.cpp
@@ -1,16 +1,0 @@
-#include "vcpch.h"
-#include "API.hpp"
-
-namespace VkApp
-{
-
-	void API::Init()
-	{
-
-	}
-
-	void API::Destroy()
-	{
-	}
-
-}
