+++ VulkanCore/src/VulkanCore/Core/Application.hpp
@@ -4,7 +4,6 @@
 #include "VulkanCore/Core/Layer.hpp"
 
 #include "VulkanCore/Core/Window.hpp"
-#include "VulkanCore/Core/API.hpp"
 
 #include "VulkanCore/ImGui/BaseImGuiLayer.hpp"
 
@@ -42,8 +41,6 @@ namespace VkApp
 		template<typename TEvent>
 		inline void DispatchEvent(TEvent e = TEvent()) { static_assert(std::is_base_of<Event, TEvent>::value); OnEvent(e); }
 
-		inline std::shared_ptr<API> GetAPIInstance() { return m_API; }
-
 		inline static Application& Get() { return *s_Instance; }
 		inline static std::filesystem::path GetWorkingDirectory() { return std::filesystem::path(s_Instance->m_AppInfo.Args[0]).parent_path(); }
 
@@ -57,8 +54,6 @@ namespace VkApp
 		AppInfo m_AppInfo;
 
 		std::unique_ptr<Window> m_Window = nullptr;
-		std::shared_ptr<API> m_API = nullptr;
-
 		bool m_Running = true;
 		bool m_Minimized = false;
