--- VulkanCore/src/VulkanCore/Core/GraphicsContext.cpp
@@ -1,933 +1,0 @@
-#include "vcpch.h"
-#include "GraphicsContext.hpp"
-
-#include "VulkanCore/Core/Logging.hpp"
-
-#define GLM_FORCE_RADIANS
-#define GLM_FORCE_DEPTH_ZERO_TO_ONE
-
-// To prevent steam errors.
-#define DISABLE_VK_LAYER_VALVE_steam_overlay_1 1
-
-// Logic for validation layers
-#ifdef VKAPP_DIST
-#define VKAPP_VALIDATION_LAYERS 0
-#else
-#define VKAPP_VALIDATION_LAYERS 1
-#endif
-
-//------------------------------------------------------------------------------------------------------------------------------------------------------
-//												External functions so we create our own function of it
-//------------------------------------------------------------------------------------------------------------------------------------------------------
-static VkResult CreateDebugUtilsMessengerEXT(VkInstance instance, const VkDebugUtilsMessengerCreateInfoEXT* pCreateInfo,
-	const VkAllocationCallbacks* pAllocator, VkDebugUtilsMessengerEXT* pDebugMessenger)
-{
-	auto func = (PFN_vkCreateDebugUtilsMessengerEXT)vkGetInstanceProcAddr(instance, "vkCreateDebugUtilsMessengerEXT");
-
-	if (func != nullptr)
-		return func(instance, pCreateInfo, pAllocator, pDebugMessenger);
-
-	return VK_ERROR_EXTENSION_NOT_PRESENT;
-}
-
-static void DestroyDebugUtilsMessengerEXT(VkInstance instance, VkDebugUtilsMessengerEXT debugMessenger, const VkAllocationCallbacks* pAllocator)
-{
-	auto func = (PFN_vkDestroyDebugUtilsMessengerEXT)vkGetInstanceProcAddr(instance, "vkDestroyDebugUtilsMessengerEXT");
-
-	if (func != nullptr)
-		func(instance, debugMessenger, pAllocator);
-}
-//------------------------------------------------------------------------------------------------------------------------------------------------------
-
-namespace VkApp
-{
-	// ---------------------------------------------------------------------------------------------------------------
-	//								Static variables and functions specfic for Cry
-	// ---------------------------------------------------------------------------------------------------------------
-
-	// Requested validation layers for vulkan debugging.
-	// This is a static variable, so it can be accessed from any function in this file
-	static const std::vector<const char*> s_RequestedValidationLayers = { "VK_LAYER_KHRONOS_validation" };
-
-	// This is needed to present to the screen
-	// This is a static variable, so it can be accessed from any function in this file
-	static const std::vector<const char*> s_RequestedDeviceExtensions = { VK_KHR_SWAPCHAIN_EXTENSION_NAME };
-
-
-
-	static VKAPI_ATTR VkBool32 VKAPI_CALL DebugCallback(
-		VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity,
-		VkDebugUtilsMessageTypeFlagsEXT messageType,
-		const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData,
-		void* pUserData)
-	{
-		if (messageSeverity >= VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT)
-			VKAPP_LOG_WARN("Validation layer: {0}", pCallbackData->pMessage);
-
-		return VK_FALSE;
-	}
-
-	static void PopulateDebugMessengerInfo(VkDebugUtilsMessengerCreateInfoEXT& createInfo)
-	{
-		createInfo = {};
-		createInfo.sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CREATE_INFO_EXT;
-		createInfo.messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT;
-		createInfo.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT;
-		createInfo.pfnUserCallback = DebugCallback;
-	}
-	// ---------------------------------------------------------------------------------------------------------------
-
-
-
-	GraphicsContext::GraphicsContext(GLFWwindow* window)
-		: m_WindowHandle(window)
-	{
-		if (window == nullptr) 
-			VKAPP_LOG_ERROR("Window handle passed in is NULL");
-	}
-
-	void GraphicsContext::Init()
-	{
-		CreateInstance();
-		SetupDebugMessenger();
-		CreateSurface();
-		GetPhysicalDevice();
-		CreateLogicalDevice();
-		CreateSwapChain();
-		CreateImageViews();
-		CreateRenderPass();
-		CreateGraphicsPipeline(); // Create a default pipeline
-		CreateFramebuffers();
-		CreateCommandPool();
-		CreateCommandBuffers();
-		CreateSyncObjects();
-	}
-
-	void GraphicsContext::Destroy()
-	{
-		vkDeviceWaitIdle(m_Info.LogicalDevice);
-
-		CleanUpSwapChain();
-
-		vkDestroyPipeline(m_Info.LogicalDevice, m_Info.GraphicsPipeline, nullptr);
-		vkDestroyPipelineLayout(m_Info.LogicalDevice, m_Info.PipelineLayout, nullptr);
-
-		vkDestroyRenderPass(m_Info.LogicalDevice, m_Info.RenderPass, nullptr);
-
-		for (size_t i = 0; i < MAX_FRAMES_IN_FLIGHT; i++)
-		{
-			vkDestroySemaphore(m_Info.LogicalDevice, m_Info.RenderFinishedSemaphores[i], nullptr);
-			vkDestroySemaphore(m_Info.LogicalDevice, m_Info.ImageAvailableSemaphores[i], nullptr);
-			vkDestroyFence(m_Info.LogicalDevice, m_Info.InFlightFences[i], nullptr);
-		}
-
-		vkDestroyCommandPool(m_Info.LogicalDevice, m_Info.CommandPool, nullptr);
-
-		vkDestroyDevice(m_Info.LogicalDevice, nullptr);
-
-		// Destroy debugger, surface & instance last
-		#if VKAPP_VALIDATION_LAYERS
-		DestroyDebugUtilsMessengerEXT(m_Info.Instance, m_Info.DebugMessenger, nullptr);
-		#endif
-
-		vkDestroySurfaceKHR(m_Info.Instance, m_Info.Surface, nullptr);
-		vkDestroyInstance(m_Info.Instance, nullptr);
-	}
-
-	void GraphicsContext::SwapBuffers()
-	{
-	}
-
-	void GraphicsContext::RecreateSwapChain()
-	{
-		// Pause if minimized
-		int width = 0, height = 0;
-		glfwGetFramebufferSize(m_WindowHandle, &width, &height);
-		while (width == 0 || height == 0) // This loop only executes if it's minimized
-		{
-			glfwGetFramebufferSize(m_WindowHandle, &width, &height);
-			glfwWaitEvents();
-		}
-
-		vkDeviceWaitIdle(m_Info.LogicalDevice);
-
-		CleanUpSwapChain();
-
-		CreateSwapChain();
-		CreateImageViews();
-		CreateFramebuffers();
-	}
-
-	void GraphicsContext::RecordCommandBuffer(VkCommandBuffer& commandBuffer, uint32_t imageIndex, const RenderInfo& info)
-	{
-		VkCommandBufferBeginInfo beginInfo = {};
-		beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-
-		if (vkBeginCommandBuffer(commandBuffer, &beginInfo) != VK_SUCCESS)
-			throw std::runtime_error("Failed to begin recording command buffer!");
-
-		VkRenderPassBeginInfo renderPassInfo = {};
-		renderPassInfo.sType = VK_STRUCTURE_TYPE_RENDER_PASS_BEGIN_INFO;
-		renderPassInfo.renderPass = m_Info.RenderPass;
-		renderPassInfo.framebuffer = m_Info.SwapChainFramebuffers[imageIndex];
-		renderPassInfo.renderArea.offset = { 0, 0 };
-		renderPassInfo.renderArea.extent = m_Info.SwapChainExtent;
-
-		VkClearValue clearColor = { {{0.0f, 0.0f, 0.0f, 1.0f}} };
-		renderPassInfo.clearValueCount = 1;
-		renderPassInfo.pClearValues = &clearColor;
-
-		vkCmdBeginRenderPass(commandBuffer, &renderPassInfo, VK_SUBPASS_CONTENTS_INLINE);
-
-		vkCmdBindPipeline(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS, m_Info.GraphicsPipeline);
-
-		VkViewport viewport = {};
-		viewport.x = 0.0f;
-		viewport.y = 0.0f;
-		viewport.width = (float)m_Info.SwapChainExtent.width;
-		viewport.height = (float)m_Info.SwapChainExtent.height;
-		viewport.minDepth = 0.0f;
-		viewport.maxDepth = 1.0f;
-		vkCmdSetViewport(commandBuffer, 0, 1, &viewport);
-
-		VkRect2D scissor = {};
-		scissor.offset = { 0, 0 };
-		scissor.extent = m_Info.SwapChainExtent;
-		vkCmdSetScissor(commandBuffer, 0, 1, &scissor);
-
-		// TODO(Jorben): Make this not be done here
-		vkCmdBindVertexBuffers(commandBuffer, 0, 1, info.VertexBuffers.data(), info.Offsets.data());
-		vkCmdBindIndexBuffer(commandBuffer, info.IndexBuffer, 0, VK_INDEX_TYPE_UINT16);
-
-		vkCmdDrawIndexed(commandBuffer, info.IndiceCount, 1, 0, 0, 0);
-
-		vkCmdEndRenderPass(commandBuffer);
-
-		if (vkEndCommandBuffer(commandBuffer) != VK_SUCCESS)
-			throw std::runtime_error("Failed to record command buffer!");
-	}
-
-	uint32_t GraphicsContext::FindMemoryType(uint32_t typeFilter, VkMemoryPropertyFlags properties)
-	{
-		VkPhysicalDeviceMemoryProperties memProperties = {};
-		vkGetPhysicalDeviceMemoryProperties(m_Info.PhysicalDevice, &memProperties);
-
-		for (uint32_t i = 0; i < memProperties.memoryTypeCount; i++) 
-		{
-			if ((typeFilter & (1 << i)) && (memProperties.memoryTypes[i].propertyFlags & properties) == properties)
-				return i;
-		}
-
-		throw std::runtime_error("Failed to find suitable memory type!");
-	}
-
-	void GraphicsContext::CreateInstance()
-	{
-		// First run a check for validation layer support
-		if (VKAPP_VALIDATION_LAYERS && !CheckValidationLayerSupport())
-			throw std::runtime_error("Validation layers requested, but not available!");
-
-		VkApplicationInfo appInfo = {};
-		appInfo.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO;
-		appInfo.pApplicationName = "CryApplication";
-		appInfo.applicationVersion = VK_MAKE_VERSION(1, 0, 0);
-		appInfo.pEngineName = "No Engine";
-		appInfo.engineVersion = VK_MAKE_VERSION(1, 0, 0);
-		appInfo.apiVersion = VK_API_VERSION_1_0;
-
-		/* DEBUG CODE: Code to check all available extensions
-		uint32_t debugExtensionCount = 0;
-		vkEnumerateInstanceExtensionProperties(nullptr, &debugExtensionCount, nullptr);
-
-		std::vector<VkExtensionProperties> debugExtensions(debugExtensionCount);
-		vkEnumerateInstanceExtensionProperties(nullptr, &debugExtensionCount, debugExtensions.data());
-
-		std::cout << "available extensions:\n";
-
-		for (const auto& extension : debugExtensions)
-			std::cout << '\t' << extension.extensionName << std::endl;
-		*/
-		auto extensions = GetRequiredExtensions();
-
-		VkInstanceCreateInfo createInfo = {};
-		createInfo.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO;
-		createInfo.pApplicationInfo = &appInfo;
-		//createInfo.flags |= VK_INSTANCE_CREATE_ENUMERATE_PORTABILITY_BIT_KHR;			// Note(Jorben): This line is needed for MacOS
-		createInfo.enabledExtensionCount = static_cast<uint32_t>(extensions.size());
-		createInfo.ppEnabledExtensionNames = extensions.data();
-
-		#if VKAPP_VALIDATION_LAYERS
-		createInfo.enabledLayerCount = static_cast<uint32_t>(s_RequestedValidationLayers.size());
-		createInfo.ppEnabledLayerNames = s_RequestedValidationLayers.data();
-		#else
-		createInfo.enabledLayerCount = 0;
-		#endif
-
-		// Note(Jorben): Setup the debug messenger also for the create instance 
-		VkDebugUtilsMessengerCreateInfoEXT debugCreateInfo = {};
-		#if VKAPP_VALIDATION_LAYERS
-		createInfo.enabledLayerCount = static_cast<uint32_t>(s_RequestedValidationLayers.size());
-		createInfo.ppEnabledLayerNames = s_RequestedValidationLayers.data();
-
-		PopulateDebugMessengerInfo(debugCreateInfo);
-		createInfo.pNext = (VkDebugUtilsMessengerCreateInfoEXT*)&debugCreateInfo;
-		#else
-		createInfo.enabledLayerCount = 0;
-		createInfo.pNext = nullptr;
-		#endif
-
-		// Create our vulkan instance
-		if (vkCreateInstance(&createInfo, nullptr, &m_Info.Instance) != VK_SUCCESS)
-		{
-			throw std::runtime_error("Failed to create Vulkan instance!");
-		}
-	}
-
-	void GraphicsContext::SetupDebugMessenger()
-	{
-		#if !(VKAPP_VALIDATION_LAYERS)
-		return;
-		#else
-		VkDebugUtilsMessengerCreateInfoEXT createInfo = {};
-		PopulateDebugMessengerInfo(createInfo);
-
-		if (CreateDebugUtilsMessengerEXT(m_Info.Instance, &createInfo, nullptr, &m_Info.DebugMessenger) != VK_SUCCESS)
-			throw std::runtime_error("Failed to set up debug messenger!");
-		#endif
-	}
-
-	void GraphicsContext::CreateSurface()
-	{
-		if (glfwCreateWindowSurface(m_Info.Instance, m_WindowHandle, nullptr, &m_Info.Surface) != VK_SUCCESS)
-			throw std::runtime_error("Failed to create window surface!");
-	}
-
-	void GraphicsContext::GetPhysicalDevice()
-	{
-		uint32_t deviceCount = 0;
-		vkEnumeratePhysicalDevices(m_Info.Instance, &deviceCount, nullptr);
-
-		if (deviceCount == 0)
-			throw std::runtime_error("Failed to find GPUs with Vulkan support!");
-
-		std::vector<VkPhysicalDevice> devices(deviceCount);
-		vkEnumeratePhysicalDevices(m_Info.Instance, &deviceCount, devices.data());
-
-		for (const auto& device : devices)
-		{
-			if (IsPhysicalDeviceSuitable(device))
-			{
-				m_Info.PhysicalDevice = device;
-				break;
-			}
-		}
-
-		// Check if no device is suitable
-		if (m_Info.PhysicalDevice == VK_NULL_HANDLE)
-			throw std::runtime_error("Failed to find a suitable GPU!");
-	}
-
-	void GraphicsContext::CreateLogicalDevice()
-	{
-		QueueFamilyIndices indices = FindQueueFamilies(m_Info.PhysicalDevice);
-
-		std::vector<VkDeviceQueueCreateInfo> queueCreateInfos;
-		std::set<uint32_t> uniqueQueueFamilies = { indices.GraphicsFamily.value(), indices.PresentFamily.value() };
-
-		float queuePriority = 1.0f;
-		for (uint32_t queueFamily : uniqueQueueFamilies)
-		{
-			VkDeviceQueueCreateInfo queueCreateInfo{};
-			queueCreateInfo.sType = VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO;
-			queueCreateInfo.queueFamilyIndex = queueFamily;
-			queueCreateInfo.queueCount = 1;
-			queueCreateInfo.pQueuePriorities = &queuePriority;
-			queueCreateInfos.push_back(queueCreateInfo);
-		}
-
-		VkPhysicalDeviceFeatures deviceFeatures = {}; // More to add when start doing interesting things with Vulkan
-
-		VkDeviceCreateInfo createInfo = {};
-		createInfo.sType = VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO;
-		createInfo.queueCreateInfoCount = static_cast<uint32_t>(queueCreateInfos.size());
-		createInfo.pQueueCreateInfos = queueCreateInfos.data();
-		createInfo.pEnabledFeatures = &deviceFeatures;
-		createInfo.enabledExtensionCount = static_cast<uint32_t>(s_RequestedDeviceExtensions.size());
-		createInfo.ppEnabledExtensionNames = s_RequestedDeviceExtensions.data();
-
-		#if VKAPP_VALIDATION_LAYERS
-		createInfo.enabledLayerCount = static_cast<uint32_t>(s_RequestedValidationLayers.size());
-		createInfo.ppEnabledLayerNames = s_RequestedValidationLayers.data();
-		#else
-		createInfo.enabledLayerCount = 0;
-		#endif
-
-		if (vkCreateDevice(m_Info.PhysicalDevice, &createInfo, nullptr, &m_Info.LogicalDevice) != VK_SUCCESS)
-			throw std::runtime_error("Failed to create logical device!");
-
-		// Retrieve the graphics & present queue handle
-		vkGetDeviceQueue(m_Info.LogicalDevice, indices.GraphicsFamily.value(), 0, &m_Info.GraphicsQueue);
-		vkGetDeviceQueue(m_Info.LogicalDevice, indices.PresentFamily.value(), 0, &m_Info.PresentQueue);
-	}
-
-	void GraphicsContext::CreateSwapChain()
-	{
-		SwapChainSupportDetails swapChainSupport = QuerySwapChainSupport(m_Info.PhysicalDevice);
-
-		VkSurfaceFormatKHR surfaceFormat = ChooseSwapSurfaceFormat(swapChainSupport.Formats);
-		VkPresentModeKHR presentMode = ChooseSwapPresentMode(swapChainSupport.PresentModes);
-		VkExtent2D extent = ChooseSwapExtent(swapChainSupport.Capabilities);
-
-		// Note(Jorben): +1 because sticking to the minimum can cause us to wait on the driver sometimes
-		uint32_t imageCount = swapChainSupport.Capabilities.minImageCount + 1;
-
-		// Making sure we don't exceed the maximum
-		if (swapChainSupport.Capabilities.maxImageCount > 0 && imageCount > swapChainSupport.Capabilities.maxImageCount)
-			imageCount = swapChainSupport.Capabilities.maxImageCount;
-
-		// Creation info
-		VkSwapchainCreateInfoKHR createInfo = {};
-		createInfo.sType = VK_STRUCTURE_TYPE_SWAPCHAIN_CREATE_INFO_KHR;
-		createInfo.surface = m_Info.Surface;
-		createInfo.minImageCount = imageCount;
-		createInfo.imageFormat = surfaceFormat.format;
-		createInfo.imageColorSpace = surfaceFormat.colorSpace;
-		createInfo.imageExtent = extent;
-		createInfo.imageArrayLayers = 1;
-		createInfo.imageUsage = VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
-
-		QueueFamilyIndices indices = FindQueueFamilies(m_Info.PhysicalDevice);
-		uint32_t queueFamilyIndices[] = { indices.GraphicsFamily.value(), indices.PresentFamily.value() };
-
-		if (indices.GraphicsFamily != indices.PresentFamily)
-		{
-			createInfo.imageSharingMode = VK_SHARING_MODE_CONCURRENT;
-			createInfo.queueFamilyIndexCount = 2;
-			createInfo.pQueueFamilyIndices = queueFamilyIndices;
-		}
-		else
-		{
-			createInfo.imageSharingMode = VK_SHARING_MODE_EXCLUSIVE;
-			createInfo.queueFamilyIndexCount = 0; // Optional
-			createInfo.pQueueFamilyIndices = nullptr; // Optional
-		}
-
-		createInfo.preTransform = swapChainSupport.Capabilities.currentTransform;
-		createInfo.compositeAlpha = VK_COMPOSITE_ALPHA_OPAQUE_BIT_KHR;
-		createInfo.presentMode = presentMode;
-		createInfo.clipped = VK_TRUE;
-
-		createInfo.oldSwapchain = VK_NULL_HANDLE;
-
-		// Creation of the swapchain
-		if (vkCreateSwapchainKHR(m_Info.LogicalDevice, &createInfo, nullptr, &m_Info.SwapChain) != VK_SUCCESS)
-			throw std::runtime_error("Failed to create swap chain!");
-
-		// Get the swapchain images
-		// Note(Jorben): We query the amount of images again, because vulkan is allowed to create a swapchain with more images.
-		vkGetSwapchainImagesKHR(m_Info.LogicalDevice, m_Info.SwapChain, &imageCount, nullptr);
-		m_Info.SwapChainImages.resize(imageCount);
-		vkGetSwapchainImagesKHR(m_Info.LogicalDevice, m_Info.SwapChain, &imageCount, m_Info.SwapChainImages.data());
-
-		// Store the format and extent for the future
-		m_Info.SwapChainImageFormat = surfaceFormat.format;
-		m_Info.SwapChainExtent = extent;
-	}
-
-	void GraphicsContext::CreateImageViews()
-	{
-		m_Info.SwapChainImageViews.resize(m_Info.SwapChainImages.size());
-
-		for (size_t i = 0; i < m_Info.SwapChainImages.size(); i++)
-		{
-			VkImageViewCreateInfo createInfo = {};
-			createInfo.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
-			createInfo.image = m_Info.SwapChainImages[i];
-			createInfo.viewType = VK_IMAGE_VIEW_TYPE_2D;
-			createInfo.format = m_Info.SwapChainImageFormat;
-			createInfo.components.r = VK_COMPONENT_SWIZZLE_IDENTITY;
-			createInfo.components.g = VK_COMPONENT_SWIZZLE_IDENTITY;
-			createInfo.components.b = VK_COMPONENT_SWIZZLE_IDENTITY;
-			createInfo.components.a = VK_COMPONENT_SWIZZLE_IDENTITY;
-			createInfo.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-			createInfo.subresourceRange.baseMipLevel = 0;
-			createInfo.subresourceRange.levelCount = 1;
-			createInfo.subresourceRange.baseArrayLayer = 0;
-			createInfo.subresourceRange.layerCount = 1;
-
-			if (vkCreateImageView(m_Info.LogicalDevice, &createInfo, nullptr, &m_Info.SwapChainImageViews[i]) != VK_SUCCESS)
-				throw std::runtime_error("Failed to create image views!");
-		}
-	}
-
-	void GraphicsContext::CreateRenderPass()
-	{
-		VkAttachmentDescription colorAttachment = {};
-		colorAttachment.format = m_Info.SwapChainImageFormat;
-		colorAttachment.samples = VK_SAMPLE_COUNT_1_BIT;
-		colorAttachment.loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-		colorAttachment.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-		colorAttachment.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-		colorAttachment.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-		colorAttachment.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-		colorAttachment.finalLayout = VK_IMAGE_LAYOUT_PRESENT_SRC_KHR;
-
-		VkAttachmentReference colorAttachmentRef = {};
-		colorAttachmentRef.attachment = 0;
-		colorAttachmentRef.layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-		VkSubpassDescription subpass = {};
-		subpass.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-		subpass.colorAttachmentCount = 1;
-		subpass.pColorAttachments = &colorAttachmentRef;
-
-		VkSubpassDependency dependency = {};
-		dependency.srcSubpass = VK_SUBPASS_EXTERNAL;
-		dependency.dstSubpass = 0;
-		dependency.srcStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-		dependency.srcAccessMask = 0;
-		dependency.dstStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-		dependency.dstAccessMask = VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-
-		VkRenderPassCreateInfo renderPassInfo = {};
-		renderPassInfo.sType = VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO;
-		renderPassInfo.attachmentCount = 1;
-		renderPassInfo.pAttachments = &colorAttachment;
-		renderPassInfo.subpassCount = 1;
-		renderPassInfo.pSubpasses = &subpass;
-		renderPassInfo.dependencyCount = 1;
-		renderPassInfo.pDependencies = &dependency;
-
-		if (vkCreateRenderPass(m_Info.LogicalDevice, &renderPassInfo, nullptr, &m_Info.RenderPass) != VK_SUCCESS)
-			throw std::runtime_error("failed to create render pass!");
-	}
-
-	void GraphicsContext::CreateGraphicsPipeline()
-	{
-		// Default triangle shader
-		auto vertShaderCode = ReadFile("assets\\shaders\\vert.spv");
-		auto fragShaderCode = ReadFile("assets\\shaders\\frag.spv");
-
-		VkShaderModule vertShaderModule = CreateShaderModule(vertShaderCode);
-		VkShaderModule fragShaderModule = CreateShaderModule(fragShaderCode);
-
-		// Vertex shader info
-		VkPipelineShaderStageCreateInfo vertShaderStageInfo = {};
-		vertShaderStageInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
-		vertShaderStageInfo.stage = VK_SHADER_STAGE_VERTEX_BIT;
-		vertShaderStageInfo.module = vertShaderModule;
-		vertShaderStageInfo.pName = "main";
-
-		// Fragment shader info
-		VkPipelineShaderStageCreateInfo fragShaderStageInfo = {};
-		fragShaderStageInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
-		fragShaderStageInfo.stage = VK_SHADER_STAGE_FRAGMENT_BIT;
-		fragShaderStageInfo.module = fragShaderModule;
-		fragShaderStageInfo.pName = "main";
-
-		// PipelineShader info
-		VkPipelineShaderStageCreateInfo shaderStages[2] = { vertShaderStageInfo, fragShaderStageInfo };
-
-		auto bindingDescription = Vertex::GetBindingDescription();
-		auto attributeDescriptions = Vertex::GetAttributeDescriptions();
-
-		VkPipelineVertexInputStateCreateInfo vertexInputInfo = {};
-		vertexInputInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO;
-		vertexInputInfo.vertexBindingDescriptionCount = 1;
-		vertexInputInfo.vertexAttributeDescriptionCount = static_cast<uint32_t>(attributeDescriptions.size());
-		vertexInputInfo.pVertexBindingDescriptions = &bindingDescription;
-		vertexInputInfo.pVertexAttributeDescriptions = attributeDescriptions.data();
-
-		VkPipelineInputAssemblyStateCreateInfo inputAssembly = {};
-		inputAssembly.sType = VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO;
-		inputAssembly.topology = VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST;
-		inputAssembly.primitiveRestartEnable = VK_FALSE;
-
-		VkPipelineViewportStateCreateInfo viewportState = {};
-		viewportState.sType = VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO;
-		viewportState.viewportCount = 1;
-		viewportState.scissorCount = 1;
-
-		VkPipelineRasterizationStateCreateInfo rasterizer = {};
-		rasterizer.sType = VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO;
-		rasterizer.depthClampEnable = VK_FALSE;
-		rasterizer.rasterizerDiscardEnable = VK_FALSE;
-		rasterizer.polygonMode = VK_POLYGON_MODE_FILL;
-		rasterizer.lineWidth = 1.0f;
-		rasterizer.cullMode = VK_CULL_MODE_BACK_BIT;
-		rasterizer.frontFace = VK_FRONT_FACE_CLOCKWISE;
-		rasterizer.depthBiasEnable = VK_FALSE;
-
-		VkPipelineMultisampleStateCreateInfo multisampling = {};
-		multisampling.sType = VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO;
-		multisampling.sampleShadingEnable = VK_FALSE;
-		multisampling.rasterizationSamples = VK_SAMPLE_COUNT_1_BIT;
-
-		VkPipelineColorBlendAttachmentState colorBlendAttachment = {};
-		colorBlendAttachment.colorWriteMask = VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
-		colorBlendAttachment.blendEnable = VK_FALSE;
-
-		VkPipelineColorBlendStateCreateInfo colorBlending = {};
-		colorBlending.sType = VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
-		colorBlending.logicOpEnable = VK_FALSE;
-		colorBlending.logicOp = VK_LOGIC_OP_COPY;
-		colorBlending.attachmentCount = 1;
-		colorBlending.pAttachments = &colorBlendAttachment;
-		colorBlending.blendConstants[0] = 0.0f;
-		colorBlending.blendConstants[1] = 0.0f;
-		colorBlending.blendConstants[2] = 0.0f;
-		colorBlending.blendConstants[3] = 0.0f;
-
-		std::vector<VkDynamicState> dynamicStates = {
-			VK_DYNAMIC_STATE_VIEWPORT,
-			VK_DYNAMIC_STATE_SCISSOR
-		};
-
-		VkPipelineDynamicStateCreateInfo dynamicState = {};
-		dynamicState.sType = VK_STRUCTURE_TYPE_PIPELINE_DYNAMIC_STATE_CREATE_INFO;
-		dynamicState.dynamicStateCount = static_cast<uint32_t>(dynamicStates.size());
-		dynamicState.pDynamicStates = dynamicStates.data();
-
-		VkPipelineLayoutCreateInfo pipelineLayoutInfo = {};
-		pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-		pipelineLayoutInfo.setLayoutCount = 0;
-		pipelineLayoutInfo.pushConstantRangeCount = 0;
-
-		if (vkCreatePipelineLayout(m_Info.LogicalDevice, &pipelineLayoutInfo, nullptr, &m_Info.PipelineLayout) != VK_SUCCESS)
-			throw std::runtime_error("failed to create pipeline layout!");
-
-		// Create the actual graphics pipeline (where we actually use the shaders and other info)
-		VkGraphicsPipelineCreateInfo pipelineInfo = {};
-		pipelineInfo.sType = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
-		pipelineInfo.stageCount = 2;
-		pipelineInfo.pStages = shaderStages;
-		pipelineInfo.pVertexInputState = &vertexInputInfo;
-		pipelineInfo.pInputAssemblyState = &inputAssembly;
-		pipelineInfo.pViewportState = &viewportState;
-		pipelineInfo.pRasterizationState = &rasterizer;
-		pipelineInfo.pMultisampleState = &multisampling;
-		pipelineInfo.pDepthStencilState = nullptr; // Optional
-		pipelineInfo.pColorBlendState = &colorBlending;
-		pipelineInfo.pDynamicState = &dynamicState;
-		pipelineInfo.layout = m_Info.PipelineLayout;
-		pipelineInfo.renderPass = m_Info.RenderPass;
-		pipelineInfo.subpass = 0;
-		pipelineInfo.basePipelineHandle = VK_NULL_HANDLE; // Optional
-		pipelineInfo.basePipelineIndex = -1; // Optional
-
-		if (vkCreateGraphicsPipelines(m_Info.LogicalDevice, VK_NULL_HANDLE, 1, &pipelineInfo, nullptr, &m_Info.GraphicsPipeline) != VK_SUCCESS)
-			throw std::runtime_error("Failed to create graphics pipeline!");
-
-		// Destroy at the end
-		vkDestroyShaderModule(m_Info.LogicalDevice, fragShaderModule, nullptr);
-		vkDestroyShaderModule(m_Info.LogicalDevice, vertShaderModule, nullptr);
-	}
-
-	void GraphicsContext::CreateFramebuffers()
-	{
-		m_Info.SwapChainFramebuffers.resize(m_Info.SwapChainImageViews.size());
-
-		for (size_t i = 0; i < m_Info.SwapChainImageViews.size(); i++)
-		{
-			VkImageView attachments[] = { m_Info.SwapChainImageViews[i] };
-
-			VkFramebufferCreateInfo framebufferInfo = {};
-			framebufferInfo.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
-			framebufferInfo.renderPass = m_Info.RenderPass;
-			framebufferInfo.attachmentCount = 1;
-			framebufferInfo.pAttachments = attachments;
-			framebufferInfo.width = m_Info.SwapChainExtent.width;
-			framebufferInfo.height = m_Info.SwapChainExtent.height;
-			framebufferInfo.layers = 1;
-
-			if (vkCreateFramebuffer(m_Info.LogicalDevice, &framebufferInfo, nullptr, &m_Info.SwapChainFramebuffers[i]) != VK_SUCCESS)
-				throw std::runtime_error("Failed to create framebuffer!");
-		}
-	}
-
-	void GraphicsContext::CreateCommandPool()
-	{
-		QueueFamilyIndices queueFamilyIndices = FindQueueFamilies(m_Info.PhysicalDevice);
-
-		VkCommandPoolCreateInfo poolInfo = {};
-		poolInfo.sType = VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO;
-		poolInfo.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
-		poolInfo.queueFamilyIndex = queueFamilyIndices.GraphicsFamily.value();
-
-		if (vkCreateCommandPool(m_Info.LogicalDevice, &poolInfo, nullptr, &m_Info.CommandPool) != VK_SUCCESS)
-			throw std::runtime_error("Failed to create command pool!");
-	}
-
-	void GraphicsContext::CreateCommandBuffers()
-	{
-		m_Info.CommandBuffers.resize(MAX_FRAMES_IN_FLIGHT);
-
-		VkCommandBufferAllocateInfo allocInfo = {};
-		allocInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-		allocInfo.commandPool = m_Info.CommandPool;
-		allocInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-		allocInfo.commandBufferCount = (uint32_t)m_Info.CommandBuffers.size();
-
-		if (vkAllocateCommandBuffers(m_Info.LogicalDevice, &allocInfo, m_Info.CommandBuffers.data()) != VK_SUCCESS)
-			throw std::runtime_error("Failed to allocate command buffers!");
-	}
-
-	void GraphicsContext::CreateSyncObjects()
-	{
-		m_Info.ImageAvailableSemaphores.resize(MAX_FRAMES_IN_FLIGHT);
-		m_Info.RenderFinishedSemaphores.resize(MAX_FRAMES_IN_FLIGHT);
-		m_Info.InFlightFences.resize(MAX_FRAMES_IN_FLIGHT);
-
-		VkSemaphoreCreateInfo semaphoreInfo = {};
-		semaphoreInfo.sType = VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO;
-
-		VkFenceCreateInfo fenceInfo = {};
-		fenceInfo.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO;
-		fenceInfo.flags = VK_FENCE_CREATE_SIGNALED_BIT;
-
-		// Create our objects
-		for (size_t i = 0; i < MAX_FRAMES_IN_FLIGHT; i++)
-		{
-			if (vkCreateSemaphore(m_Info.LogicalDevice, &semaphoreInfo, nullptr, &m_Info.ImageAvailableSemaphores[i]) != VK_SUCCESS ||
-				vkCreateSemaphore(m_Info.LogicalDevice, &semaphoreInfo, nullptr, &m_Info.RenderFinishedSemaphores[i]) != VK_SUCCESS ||
-				vkCreateFence(m_Info.LogicalDevice, &fenceInfo, nullptr, &m_Info.InFlightFences[i]) != VK_SUCCESS)
-			{
-
-				throw std::runtime_error("Failed to create synchronization objects for a frame!");
-			}
-		}
-	}
-
-	void GraphicsContext::CleanUpSwapChain()
-	{
-		for (size_t i = 0; i < m_Info.SwapChainFramebuffers.size(); i++)
-			vkDestroyFramebuffer(m_Info.LogicalDevice, m_Info.SwapChainFramebuffers[i], nullptr);
-
-		for (size_t i = 0; i < m_Info.SwapChainImageViews.size(); i++)
-			vkDestroyImageView(m_Info.LogicalDevice, m_Info.SwapChainImageViews[i], nullptr);
-
-		vkDestroySwapchainKHR(m_Info.LogicalDevice, m_Info.SwapChain, nullptr);
-	}
-
-	bool GraphicsContext::CheckValidationLayerSupport()
-	{
-		uint32_t layerCount;
-		vkEnumerateInstanceLayerProperties(&layerCount, nullptr);
-
-		std::vector<VkLayerProperties> availableLayers(layerCount);
-		vkEnumerateInstanceLayerProperties(&layerCount, availableLayers.data());
-
-		// Check if all requested layers are actually accessible 
-		for (const char* layerName : s_RequestedValidationLayers)
-		{
-			bool layerFound = false;
-
-			for (const auto& layerProperties : availableLayers)
-			{
-				if (strcmp(layerName, layerProperties.layerName) == 0)
-				{
-					layerFound = true;
-					break;
-				}
-			}
-
-			if (!layerFound)
-				return false;
-		}
-
-		return true;
-	}
-
-	std::vector<const char*> GraphicsContext::GetRequiredExtensions()
-	{
-		uint32_t glfwExtensionCount = 0;
-		const char** glfwExtensions;
-		glfwExtensions = glfwGetRequiredInstanceExtensions(&glfwExtensionCount);
-
-		std::vector<const char*> extensions(glfwExtensions, glfwExtensions + glfwExtensionCount);
-
-		#if VKAPP_VALIDATION_LAYERS
-		extensions.push_back(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
-		#endif
-
-		//extensions.push_back(VK_KHR_PORTABILITY_ENUMERATION_EXTENSION_NAME); // Note(Jorben): This line is needed for MacOS
-
-		return extensions;
-	}
-
-	bool GraphicsContext::IsPhysicalDeviceSuitable(const VkPhysicalDevice& device)
-	{
-		QueueFamilyIndices indices = FindQueueFamilies(device);
-
-		bool extensionsSupported = CheckDeviceExtensionSupport(device);
-
-		bool swapChainAdequate = false;
-		if (extensionsSupported)
-		{
-			SwapChainSupportDetails swapChainSupport = QuerySwapChainSupport(device);
-			swapChainAdequate = !swapChainSupport.Formats.empty() && !swapChainSupport.PresentModes.empty();
-		}
-
-		return indices.IsComplete() && extensionsSupported && swapChainAdequate;
-	}
-
-	bool GraphicsContext::CheckDeviceExtensionSupport(const VkPhysicalDevice& device)
-	{
-		uint32_t extensionCount;
-		vkEnumerateDeviceExtensionProperties(device, nullptr, &extensionCount, nullptr);
-
-		std::vector<VkExtensionProperties> availableExtensions(extensionCount);
-		vkEnumerateDeviceExtensionProperties(device, nullptr, &extensionCount, availableExtensions.data());
-
-		std::set<std::string> requiredExtensions(s_RequestedDeviceExtensions.begin(), s_RequestedDeviceExtensions.end());
-
-		for (const auto& extension : availableExtensions)
-			requiredExtensions.erase(extension.extensionName);
-
-		// It's empty if all the required extensions are available
-		return requiredExtensions.empty();
-	}
-
-	QueueFamilyIndices GraphicsContext::FindQueueFamilies(const VkPhysicalDevice& device)
-	{
-		QueueFamilyIndices indices;
-
-		uint32_t queueFamilyCount = 0;
-		vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, nullptr);
-
-		std::vector<VkQueueFamilyProperties> queueFamilies(queueFamilyCount);
-		vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, queueFamilies.data());
-
-		int32_t i = 0;
-		for (const auto& queueFamily : queueFamilies)
-		{
-			// Early exit check
-			if (indices.IsComplete())
-				break;
-
-			if (queueFamily.queueFlags & VK_QUEUE_GRAPHICS_BIT)
-				indices.GraphicsFamily = i;
-
-			VkBool32 presentSupport = false;
-			vkGetPhysicalDeviceSurfaceSupportKHR(device, i, m_Info.Surface, &presentSupport);
-			if (presentSupport)
-				indices.PresentFamily = i;
-
-			i++;
-		}
-
-		return indices;
-	}
-
-	SwapChainSupportDetails GraphicsContext::QuerySwapChainSupport(const VkPhysicalDevice& device)
-	{
-		SwapChainSupportDetails details;
-
-		// Capabilities
-		vkGetPhysicalDeviceSurfaceCapabilitiesKHR(device, m_Info.Surface, &details.Capabilities);
-
-		// Formats
-		uint32_t formatCount;
-		vkGetPhysicalDeviceSurfaceFormatsKHR(device, m_Info.Surface, &formatCount, nullptr);
-
-		if (formatCount != 0)
-		{
-			details.Formats.resize((size_t)formatCount);
-			vkGetPhysicalDeviceSurfaceFormatsKHR(device, m_Info.Surface, &formatCount, details.Formats.data());
-		}
-
-		// Presentation modes
-		uint32_t presentModeCount;
-		vkGetPhysicalDeviceSurfacePresentModesKHR(device, m_Info.Surface, &presentModeCount, nullptr);
-
-		if (presentModeCount != 0)
-		{
-			details.PresentModes.resize((size_t)presentModeCount);
-			vkGetPhysicalDeviceSurfacePresentModesKHR(device, m_Info.Surface, &presentModeCount, details.PresentModes.data());
-		}
-
-		return details;
-	}
-
-	VkShaderModule GraphicsContext::CreateShaderModule(const std::vector<char>& code)
-	{
-		VkShaderModuleCreateInfo createInfo = {};
-		createInfo.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
-		createInfo.codeSize = code.size();
-		createInfo.pCode = reinterpret_cast<const uint32_t*>(code.data());
-
-		VkShaderModule shaderModule;
-		if (vkCreateShaderModule(m_Info.LogicalDevice, &createInfo, nullptr, &shaderModule) != VK_SUCCESS)
-			throw std::runtime_error("Failed to create shader module!");
-
-		return shaderModule;
-	}
-
-	VkSurfaceFormatKHR GraphicsContext::ChooseSwapSurfaceFormat(const std::vector<VkSurfaceFormatKHR>& availableFormats)
-	{
-		for (const auto& availableFormat : availableFormats)
-		{
-			if (availableFormat.format == VK_FORMAT_B8G8R8A8_SRGB && availableFormat.colorSpace == VK_COLOR_SPACE_SRGB_NONLINEAR_KHR)
-				return availableFormat;
-		}
-
-		// If nothing 100% satisfies our needs it okay to just go with the first one.
-		return availableFormats[0];
-	}
-
-	VkPresentModeKHR GraphicsContext::ChooseSwapPresentMode(const std::vector<VkPresentModeKHR>& availablePresentModes)
-	{
-		for (const auto& availablePresentMode : availablePresentModes)
-		{
-			if (availablePresentMode == VK_PRESENT_MODE_MAILBOX_KHR)
-				return availablePresentMode;
-		}
-
-		return VK_PRESENT_MODE_FIFO_KHR;
-	}
-
-	// Note(Jorben): Somehow glm is included and they have a define for max(a, b) but it interferes with numeric_limits.
-	#undef max(a, b) 
-
-	VkExtent2D GraphicsContext::ChooseSwapExtent(const VkSurfaceCapabilitiesKHR& capabilities)
-	{
-		if (capabilities.currentExtent.width != std::numeric_limits<uint32_t>::max())
-			return capabilities.currentExtent;
-		else
-		{
-			int width, height;
-			glfwGetFramebufferSize(m_WindowHandle, &width, &height);
-
-			VkExtent2D actualExtent =
-			{
-				static_cast<uint32_t>(width),
-				static_cast<uint32_t>(height)
-			};
-
-			actualExtent.width = std::clamp(actualExtent.width, capabilities.minImageExtent.width, capabilities.maxImageExtent.width);
-			actualExtent.height = std::clamp(actualExtent.height, capabilities.minImageExtent.height, capabilities.maxImageExtent.height);
-
-			return actualExtent;
-		}
-	}
-
-	std::vector<char> GraphicsContext::ReadFile(const std::filesystem::path& filename)
-	{
-		std::ifstream file(filename, std::ios::ate | std::ios::binary);
-
-		if (!file.is_open() || !file.good())
-			throw std::runtime_error("Failed to open file!");
-
-		size_t fileSize = (size_t)file.tellg();
-		std::vector<char> buffer(fileSize);
-
-		file.seekg(0);
-		file.read(buffer.data(), fileSize);
-
-		file.close();
-
-		return buffer;
-	}
-
-
-}
