+++ VulkanCore/src/VulkanCore/Renderer/GraphicsPipelineManager.cpp
@@ -1,0 +1,205 @@
+#include "vcpch.h"
+#include "GraphicsPipelineManager.hpp"
+
+#define GLFW_INCLUDE_VULKAN
+#include <GLFW/glfw3.h>
+
+#include "VulkanCore/Core/Logging.hpp"
+
+#include "VulkanCore/Core/Application.hpp"
+
+#include "VulkanCore/Renderer/InstanceManager.hpp" // For the retrieval of the logical device
+#include "VulkanCore/Renderer/SwapChainManager.hpp"
+
+namespace VkApp
+{
+
+	// ===================================
+	// ------------ Static ---------------
+	// ===================================
+	GraphicsPipelineManager* GraphicsPipelineManager::s_Instance = nullptr;
+
+	static InstanceManager* s_InstanceManager = nullptr;
+
+	// ===================================
+	// ------------ Public ---------------
+	// ===================================
+	GraphicsPipelineManager::GraphicsPipelineManager()
+	{
+		// Retrieve the InstanceManager, so we can use it in our code
+		s_Instance = this;
+		s_InstanceManager = InstanceManager::Get();
+
+		// Note(Jorben): Creates a default pipeline for drawing triangles using the custom Vertex struct
+		CreateGraphicsPipeline();
+	}
+
+	void GraphicsPipelineManager::Destroy()
+	{
+		vkDestroyPipeline(s_InstanceManager->m_Device, m_GraphicsPipeline, nullptr);
+		vkDestroyPipelineLayout(s_InstanceManager->m_Device, m_PipelineLayout, nullptr);
+
+		s_InstanceManager = nullptr;
+		s_Instance = nullptr;
+	}
+
+	std::vector<char> GraphicsPipelineManager::ReadFile(const std::filesystem::path& path)
+	{
+		std::ifstream file(path, std::ios::ate | std::ios::binary);
+
+		if (!file.is_open() || !file.good())
+			VKAPP_LOG_ERROR("Failed to open file!");
+
+		size_t fileSize = (size_t)file.tellg();
+		std::vector<char> buffer(fileSize);
+
+		file.seekg(0);
+		file.read(buffer.data(), fileSize);
+
+		file.close();
+		return buffer;
+	}
+
+	VkShaderModule GraphicsPipelineManager::CreateShaderModule(const std::vector<char>& data)
+	{
+		VkShaderModuleCreateInfo createInfo = {};
+		createInfo.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
+		createInfo.codeSize = data.size();
+		createInfo.pCode = reinterpret_cast<const uint32_t*>(data.data());
+
+		VkShaderModule shaderModule = {};
+		if (vkCreateShaderModule(s_InstanceManager->m_Device, &createInfo, nullptr, &shaderModule) != VK_SUCCESS)
+			VKAPP_LOG_ERROR("Failed to create shader module!");
+
+		return shaderModule;
+	}
+
+	// ===================================
+	// -------- Initialization -----------
+	// ===================================
+	void GraphicsPipelineManager::CreateGraphicsPipeline()
+	{
+		// Note(Jorben): This is reading a default triangle shader
+		auto vertShaderCode = ReadFile("assets\\shaders\\vert.spv");
+		auto fragShaderCode = ReadFile("assets\\shaders\\frag.spv");
+
+		VkShaderModule vertShaderModule = CreateShaderModule(vertShaderCode);
+		VkShaderModule fragShaderModule = CreateShaderModule(fragShaderCode);
+
+		// Vertex shader info
+		VkPipelineShaderStageCreateInfo vertShaderStageInfo = {};
+		vertShaderStageInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
+		vertShaderStageInfo.stage = VK_SHADER_STAGE_VERTEX_BIT;
+		vertShaderStageInfo.module = vertShaderModule;
+		vertShaderStageInfo.pName = "main";
+
+		// Fragment shader info
+		VkPipelineShaderStageCreateInfo fragShaderStageInfo = {};
+		fragShaderStageInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
+		fragShaderStageInfo.stage = VK_SHADER_STAGE_FRAGMENT_BIT;
+		fragShaderStageInfo.module = fragShaderModule;
+		fragShaderStageInfo.pName = "main";
+
+		// PipelineShader info
+		VkPipelineShaderStageCreateInfo shaderStages[2] = { vertShaderStageInfo, fragShaderStageInfo };
+
+		auto bindingDescription = Vertex::GetBindingDescription();
+		auto attributeDescriptions = Vertex::GetAttributeDescriptions();
+
+		VkPipelineVertexInputStateCreateInfo vertexInputInfo = {};
+		vertexInputInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO;
+		vertexInputInfo.vertexBindingDescriptionCount = 1;
+		vertexInputInfo.vertexAttributeDescriptionCount = static_cast<uint32_t>(attributeDescriptions.size());
+		vertexInputInfo.pVertexBindingDescriptions = &bindingDescription;
+		vertexInputInfo.pVertexAttributeDescriptions = attributeDescriptions.data();
+
+		VkPipelineInputAssemblyStateCreateInfo inputAssembly = {};
+		inputAssembly.sType = VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO;
+		inputAssembly.topology = VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST;
+		inputAssembly.primitiveRestartEnable = VK_FALSE;
+
+		VkPipelineViewportStateCreateInfo viewportState = {};
+		viewportState.sType = VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO;
+		viewportState.viewportCount = 1;
+		viewportState.scissorCount = 1;
+
+		VkPipelineRasterizationStateCreateInfo rasterizer = {};
+		rasterizer.sType = VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO;
+		rasterizer.depthClampEnable = VK_FALSE;
+		rasterizer.rasterizerDiscardEnable = VK_FALSE;
+		rasterizer.polygonMode = VK_POLYGON_MODE_FILL;
+		rasterizer.lineWidth = 1.0f;
+		rasterizer.cullMode = VK_CULL_MODE_BACK_BIT;
+		rasterizer.frontFace = VK_FRONT_FACE_CLOCKWISE;
+		rasterizer.depthBiasEnable = VK_FALSE;
+
+		VkPipelineMultisampleStateCreateInfo multisampling = {};
+		multisampling.sType = VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO;
+		multisampling.sampleShadingEnable = VK_FALSE;
+		multisampling.rasterizationSamples = VK_SAMPLE_COUNT_1_BIT;
+
+		VkPipelineColorBlendAttachmentState colorBlendAttachment = {};
+		colorBlendAttachment.colorWriteMask = VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
+		colorBlendAttachment.blendEnable = VK_FALSE;
+
+		VkPipelineColorBlendStateCreateInfo colorBlending = {};
+		colorBlending.sType = VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
+		colorBlending.logicOpEnable = VK_FALSE;
+		colorBlending.logicOp = VK_LOGIC_OP_COPY;
+		colorBlending.attachmentCount = 1;
+		colorBlending.pAttachments = &colorBlendAttachment;
+		colorBlending.blendConstants[0] = 0.0f;
+		colorBlending.blendConstants[1] = 0.0f;
+		colorBlending.blendConstants[2] = 0.0f;
+		colorBlending.blendConstants[3] = 0.0f;
+
+		std::vector<VkDynamicState> dynamicStates = {
+			VK_DYNAMIC_STATE_VIEWPORT,
+			VK_DYNAMIC_STATE_SCISSOR
+		};
+
+		VkPipelineDynamicStateCreateInfo dynamicState = {};
+		dynamicState.sType = VK_STRUCTURE_TYPE_PIPELINE_DYNAMIC_STATE_CREATE_INFO;
+		dynamicState.dynamicStateCount = static_cast<uint32_t>(dynamicStates.size());
+		dynamicState.pDynamicStates = dynamicStates.data();
+
+		VkPipelineLayoutCreateInfo pipelineLayoutInfo = {};
+		pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
+		pipelineLayoutInfo.setLayoutCount = 0;
+		pipelineLayoutInfo.pushConstantRangeCount = 0;
+
+		if (vkCreatePipelineLayout(s_InstanceManager->m_Device, &pipelineLayoutInfo, nullptr, &m_PipelineLayout) != VK_SUCCESS)
+			VKAPP_LOG_ERROR("Failed to create pipeline layout!");
+
+		// Create the actual graphics pipeline (where we actually use the shaders and other info)
+		VkGraphicsPipelineCreateInfo pipelineInfo = {};
+		pipelineInfo.sType = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
+		pipelineInfo.stageCount = 2;
+		pipelineInfo.pStages = shaderStages;
+		pipelineInfo.pVertexInputState = &vertexInputInfo;
+		pipelineInfo.pInputAssemblyState = &inputAssembly;
+		pipelineInfo.pViewportState = &viewportState;
+		pipelineInfo.pRasterizationState = &rasterizer;
+		pipelineInfo.pMultisampleState = &multisampling;
+		pipelineInfo.pDepthStencilState = nullptr; // Optional
+		pipelineInfo.pColorBlendState = &colorBlending;
+		pipelineInfo.pDynamicState = &dynamicState;
+		pipelineInfo.layout = m_PipelineLayout;
+		pipelineInfo.renderPass = SwapChainManager::Get()->GetRenderPass();
+		pipelineInfo.subpass = 0;
+		pipelineInfo.basePipelineHandle = VK_NULL_HANDLE; // Optional
+		pipelineInfo.basePipelineIndex = -1; // Optional
+
+		if (vkCreateGraphicsPipelines(s_InstanceManager->m_Device, VK_NULL_HANDLE, 1, &pipelineInfo, nullptr, &m_GraphicsPipeline) != VK_SUCCESS)
+			VKAPP_LOG_ERROR("Failed to create graphics pipeline!");
+
+		// Destroy at the end
+		vkDestroyShaderModule(s_InstanceManager->m_Device, fragShaderModule, nullptr);
+		vkDestroyShaderModule(s_InstanceManager->m_Device, vertShaderModule, nullptr);
+	}
+
+	// ===================================
+	// ------------ Helper ---------------
+	// ===================================
+
+}
